[FEATURE] Transactions/ search done with paging

+++ computer-database/src/main/java/fr/ebiz/computer_database/service/ComputerService.java
@@ -1,5 +1,7 @@
 package fr.ebiz.computer_database.service;
 
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.util.List;
 
 import fr.ebiz.computer_database.exceptions.DAOException;
@@ -8,6 +10,8 @@ import fr.ebiz.computer_database.mapper.ComputerMapper;
 import fr.ebiz.computer_database.model.Computer;
 import fr.ebiz.computer_database.model.ComputerDTO;
 import fr.ebiz.computer_database.persistence.ComputerDAO;
+import fr.ebiz.computer_database.persistence.Persistence;
+import fr.ebiz.computer_database.persistence.Transaction;
 
 /**
  * @author ckeita
@@ -24,7 +28,16 @@ public class ComputerService {
      * @throws DAOException .
      */
     public ComputerDTO findComputerById(int id) throws DateException, DAOException {
-        return computerMapper.getById(computerDAO.findById(id));
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        ComputerDTO computerDTO = computerMapper.getById(computerDAO.findById(id));
+        try {
+            connection.close();
+            Transaction.removeTransaction();
+            return computerDTO;
+        } catch (SQLException e) {
+            throw new DAOException("[findComputerById] " + e.getMessage());
+        }
     }
 
     /**
@@ -32,7 +45,16 @@ public class ComputerService {
      * @throws DAOException .
      */
     public int getNumberOfComputers() throws DAOException {
-        return computerDAO.getNumberOfComputers();
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        int nbComputers = computerDAO.getNumberOfComputers();
+        try {
+            connection.close();
+            Transaction.removeTransaction();
+            return nbComputers;
+        } catch (SQLException e) {
+            throw new DAOException("[getNumberOfComputers] " + e.getMessage());
+        }
     }
 
     /**
@@ -41,7 +63,16 @@ public class ComputerService {
      * @throws DAOException .
      */
     public int getNumberOfSearchedComputers(String name) throws DAOException {
-        return computerDAO.getNumberOfSearchedComputers(name);
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        int nbFoundComputers = computerDAO.getNumberOfSearchedComputers(name);
+        try {
+            connection.close();
+            Transaction.removeTransaction();
+            return nbFoundComputers;
+        } catch (SQLException e) {
+            throw new DAOException("[getNumberOfSearchedComputers] " + e.getMessage());
+        }
     }
 
     /**
@@ -52,7 +83,16 @@ public class ComputerService {
      * @throws DateException .
      */
     public List<ComputerDTO> findComputersByLimit(int offset, int max) throws DateException, DAOException {
-        return computerMapper.getByPage(computerDAO.findByLimit(offset, max));
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        List<ComputerDTO> computers = computerMapper.getByPage(computerDAO.findByLimit(offset, max));
+        try {
+            connection.close();
+            Transaction.removeTransaction();
+            return computers;
+        } catch (SQLException e) {
+            throw new DAOException("[findComputersByLimit] " + e.getMessage());
+        }
     }
 
     /**
@@ -64,7 +104,16 @@ public class ComputerService {
      * @throws DateException .
      */
     public List<ComputerDTO> searchComputers(String name, int offset, int max) throws DateException, DAOException {
-        return computerMapper.getByPage(computerDAO.searchByLimit(name, offset, max));
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        List<ComputerDTO> computers = computerMapper.getByPage(computerDAO.searchByLimit(name, offset, max));
+        try {
+            connection.close();
+            Transaction.removeTransaction();
+            return computers;
+        } catch (SQLException e) {
+            throw new DAOException("[findComputersByLimit] " + e.getMessage());
+        }
     }
 
     /**
@@ -72,7 +121,28 @@ public class ComputerService {
      * @throws DAOException .
      */
     public void createComputer(Computer comp) throws DAOException {
-        computerDAO.create(comp);
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        if (connection != null) {
+            try {
+                computerDAO.create(comp);
+                connection.commit();
+            } catch (SQLException e) {
+                try {
+                    connection.rollback();
+                } catch (SQLException e1) {
+                    throw new DAOException("[createComputer] " + e1.getMessage());
+                }
+                throw new DAOException("[createComputer] " + e.getMessage());
+            } finally {
+                try {
+                    connection.close();
+                    Transaction.removeTransaction();
+                } catch (SQLException e) {
+                    throw new DAOException("[createComputer] " + e.getMessage());
+                }
+            }
+        }
     }
 
     /**
@@ -80,7 +150,28 @@ public class ComputerService {
      * @exception DAOException .
      */
     public void deleteComputer(int compId) throws DAOException {
-        computerDAO.delete(compId);
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        if (connection != null) {
+            try {
+                computerDAO.delete(compId);
+                connection.commit();
+            } catch (SQLException e) {
+                try {
+                    connection.rollback();
+                } catch (SQLException e1) {
+                    throw new DAOException("[deleteComputer] " + e1.getMessage());
+                }
+                throw new DAOException("[deleteComputer] " + e.getMessage());
+            } finally {
+                try {
+                    connection.close();
+                    Transaction.removeTransaction();
+                } catch (SQLException e) {
+                    throw new DAOException("[deleteComputer] " + e.getMessage());
+                }
+            }
+        }
     }
 
     /**
@@ -88,16 +179,27 @@ public class ComputerService {
      * @exception DAOException .
      */
     public void updateComputer(Computer comp) throws DAOException {
-        computerDAO.update(comp);
-    }
-
-    /**
-     * @param name to search
-     * @return the list of computers DTO
-     * @throws DAOException .
-     * @throws DateException .
-     */
-    public List<ComputerDTO> searchComputer(String name) throws DateException, DAOException {
-        return computerMapper.getByPage(computerDAO.search(name));
+        Connection connection = Persistence.getInstance().getConnection();
+        Transaction.setTransaction(connection);
+        if (connection != null) {
+            try {
+                computerDAO.update(comp);
+                connection.commit();
+            } catch (SQLException e) {
+                try {
+                    connection.rollback();
+                } catch (SQLException e1) {
+                    throw new DAOException("[updateComputer] " + e1.getMessage());
+                }
+                throw new DAOException("[updateComputer] " + e.getMessage());
+            } finally {
+                try {
+                    connection.close();
+                    Transaction.removeTransaction();
+                } catch (SQLException e) {
+                    throw new DAOException("[updateComputer] " + e.getMessage());
+                }
+            }
+        }
     }
 }
